+++ src/api.rs
@@ -1,4 +1,4 @@
-use crate::common::{config_path, validate_password};
+use crate::common::{self, config_path, validate_password};
 use crate::middleware::TOKEN_EXPIRE_HOURS;
 use crate::socket_client::*;
 use actix_files::NamedFile;
@@ -12,7 +12,6 @@ use argon2::{
 };
 use jwt_simple::prelude::*;
 use log::{debug, error};
-use reqwest::blocking::get;
 use serde::Deserialize;
 use serde_repr::{Deserialize_repr, Serialize_repr};
 use std::{
@@ -80,18 +79,6 @@ pub enum FactoryResetMode {
     Mode2 = 2,
     Mode3 = 3,
     Mode4 = 4,
-}
-
-#[derive(Deserialize)]
-pub struct RealmInfo {
-    public_key: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-struct TokenClaims {
-    roles: Option<Vec<String>>,
-    tenant_list: Option<Vec<String>>,
-    fleet_list: Option<Vec<String>>,
 }
 
 #[derive(Clone)]
@@ -304,9 +291,12 @@ impl Api {
     pub async fn validate_portal_token(body: String, config: web::Data<Api>) -> impl Responder {
         debug!("validate_portal_token() called");
 
-        if let Err(e) =
-            Api::validate_token_and_claims(&body, &config.keycloak_public_key_url, &config.tenant)
-                .await
+        if let Err(e) = common::validate_token_and_claims(
+            &body,
+            &config.keycloak_public_key_url,
+            &config.tenant,
+        )
+        .await
         {
             error!("validate_portal_token() failed: {e:#}");
             return HttpResponse::Unauthorized().finish();
@@ -400,76 +390,5 @@ impl Api {
         }
 
         Ok(token)
-    }
-
-    async fn get_keycloak_realm_public_key(
-        keycloak_public_key_url: &str,
-    ) -> Result<RS256PublicKey> {
-        let resp = get(keycloak_public_key_url)
-            .context("failed to fetch from url")?
-            .json::<RealmInfo>()
-            .context("failed to parse realm info")?;
-
-        let base64_key = &resp.public_key;
-
-        let mut pem = String::from("-----BEGIN PUBLIC KEY-----\n");
-        for chunk in base64_key.as_bytes().chunks(64) {
-            pem.push_str(&String::from_utf8_lossy(chunk));
-            pem.push('\n');
-        }
-        pem.push_str("-----END PUBLIC KEY-----\n");
-
-        let public_key = RS256PublicKey::from_pem(&pem).context("failed to create pem")?;
-
-        Ok(public_key)
-    }
-
-    async fn validate_token_and_claims(
-        token: &str,
-        keycloak_public_key_url: &str,
-        tenant: &String,
-    ) -> Result<()> {
-        let Ok(pub_key) = Api::get_keycloak_realm_public_key(keycloak_public_key_url).await else {
-            bail!("failed to get public key");
-        };
-
-        let Ok(claims) = pub_key.verify_token::<TokenClaims>(token, None) else {
-            bail!("failed to verify token");
-        };
-
-        let Some(tenant_list) = &claims.custom.tenant_list else {
-            bail!("user has no tenant list");
-        };
-
-        if !tenant_list.contains(tenant) {
-            bail!("user has no permission to set password");
-        }
-
-        let Some(roles) = &claims.custom.roles else {
-            bail!("user has no roles");
-        };
-
-        if roles.contains(&String::from("FleetAdministrator")) {
-            return Ok(());
-        }
-
-        if roles.contains(&String::from("FleetObserver")) {
-            bail!("user has no permission to set password");
-        }
-
-        if roles.contains(&String::from("FleetOperator")) {
-            let Some(fleet_list) = &claims.custom.fleet_list else {
-                bail!("user has no permission on this fleet");
-            };
-
-            //TODO: Check if fleet is available and compare with fleet from ods
-            if !fleet_list.contains(&String::from("123")) {
-                bail!("user has no permission on this fleet");
-            } else {
-                return Ok(());
-            }
-        }
-
-        Err(anyhow!("user has no permission to set password"))
     }
 }
